relocate Bodu.Core-only contracts and KATs out of Bodu.Test
The single-consumer rule was applied: each contract / KAT below was used
only by Bodu.Core.Test, so it moves into Bodu.Core/test/ alongside its
consumers. Bodu.Test keeps the assertion helpers, marker interfaces,
stream mocks, generic KAT primitives, and the one multi-consumer
contract (ParseFormatContractTests<T>).

Contracts moved to Bodu.Core/test/Contracts/ (namespace
Bodu.Collections.Generic.Contracts to match existing subclass namespace):

- CollectionContractTests<TCollection, TItem>
- ReadOnlyCollectionContractTests<TCollection, TItem>
- SetContractTests<TSet, TItem>
- EnumeratorContractTests<TEnumerable, TItem>
- DebugViewContractTests<TCollection>
- NonGenericCollectionContractTests<TCollection>

KATs moved to Bodu.Core/test/Buffers/ (namespace Bodu.Buffers to match
PooledBufferBuilderKatTests):

- BufferCapacityKat
- BufferWriteKat

Subclass `using` updates:
- Bodu.Collections.Generic.Contracts subclasses: removed unused
  `using Bodu.Test.Contracts;` (now in same namespace as the bases).
- Bodu.Collections.Generic.Concurrent.Contracts subclasses: rewrote
  `using Bodu.Test.Contracts;` to `using Bodu.Collections.Generic.Contracts;`.
- PooledBufferBuilderKatTests: removed unused `using Bodu.Test.Kat;`.

Verified: dotnet build Bodu.Core/test passes with 0 errors;
dotnet test --settings bvt.runsettings passes 24724/24724.
