+++ Bodu.Core/test/Buffers/PooledBufferBuilderKatTests.cs
@@ -4,8 +4,6 @@
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
-using Bodu.Test.Kat;
-
 namespace Bodu.Buffers;
 
 /// <summary>
