+++ Bodu.Core/test/Collections.Generic.Concurrent/ConcurrentHashSetDebugViewContractTests.cs
@@ -4,7 +4,7 @@
 // </copyright>
 // ---------------------------------------------------------------------------------------------------------------
 
-using Bodu.Test.Contracts;
+using Bodu.Collections.Generic.Contracts;
 
 namespace Bodu.Collections.Generic.Concurrent.Contracts;
